Moving TwintSpy to Test target
