+++ AdyenActions/Components/SDK/Twint+Injectable.swift
@@ -1,0 +1,36 @@
+//
+// Copyright (c) 2024 Adyen N.V.
+//
+// This file is open source and available under the MIT license. See the LICENSE file for more info.
+//
+
+import Foundation
+#if canImport(TwintSDK)
+    import TwintSDK
+#endif
+
+#if canImport(TwintSDK)
+    extension Twint {
+        @objc func fetchInstalledAppConfigurations(completion: @escaping ([TWAppConfiguration]) -> Void) {
+            Twint.fetchInstalledAppConfigurations { configurations in
+                completion(configurations ?? [])
+            }
+        }
+        
+        @objc func pay(withCode code: String, appConfiguration: TWAppConfiguration, callback: String) -> Error? {
+            Twint.pay(withCode: code, appConfiguration: appConfiguration, callback: callback)
+        }
+        
+        @objc func controller(
+            for installedAppConfigurations: [TWAppConfiguration],
+            selectionHandler: @escaping (TWAppConfiguration?) -> Void,
+            cancelHandler: @escaping () -> Void
+        ) -> UIAlertController? {
+            Twint.controller(
+                for: installedAppConfigurations.map { $0 },
+                selectedConfigurationHandler: { selectionHandler($0) },
+                cancelHandler: { cancelHandler() }
+            )
+        }
+    }
+#endif
